reorg facts into a new module
